refactoring photo for PhotoCommon
