+++ src/components/Photo/PhotoCommon/PhotoCommon.module.scss
@@ -30,3 +30,21 @@
 .photoCommonSlide {
   margin-right: 15px;
 }
+
+
+// .sliderContainer {
+//   width: 100%;
+//   max-width: 1200px;
+//   margin: 0 auto;
+// }
+
+// .slide {
+//   cursor: pointer;
+// }
+
+// .photo {
+//   width: 100%;
+//   height: auto;
+//   border-radius: 8px;
+//   box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+// }
