+++ src/data/photoDataChalet.js
@@ -1,0 +1,5 @@
+const photoDataChalet = [
+    { id: 1, src: `${process.env.PUBLIC_URL}/images/chaletPhoto/chalet.jpg`, alt: 'Шале' },
+]
+
+export default photoDataChalet;
