+++ src/data/photoDataCommon.js
@@ -1,0 +1,67 @@
+const photoDataCommon = [
+    {
+        id: 1, src: `${process.env.PUBLIC_URL}/images/commonPhoto/altanka1.jpg`, alt: 'Альтанка'
+    },
+    {
+        id: 2, src: `${process.env.PUBLIC_URL}/images/commonPhoto/altanka2.jpg`, alt: 'Альтанка'
+    },
+    {
+        id: 3, src: `${process.env.PUBLIC_URL}/images/commonPhoto/altanka3.jpg`, alt: 'Альтанка'
+    },
+    {
+        id: 4, src: `${process.env.PUBLIC_URL}/images/commonPhoto/altanka2_5.jpg`, alt: 'Альтанка'
+    },
+    {
+        id: 5, src: `${process.env.PUBLIC_URL}/images/commonPhoto/alley3.jpg`, alt: 'Алея'
+    },
+    {
+        id: 6, src: `${process.env.PUBLIC_URL}/images/commonPhoto/alley4.jpg`, alt: 'Алея'
+    },
+    {
+        id: 7, src: `${process.env.PUBLIC_URL}/images/commonPhoto/alley5.jpg`, alt: 'Алея'
+    },
+    {
+        id: 8, src: `${process.env.PUBLIC_URL}/images/commonPhoto/alley6.jpg`, alt: 'Алея'
+    },
+    {
+        id: 9, src: `${process.env.PUBLIC_URL}/images/commonPhoto/alley7.jpg`, alt: 'Алея'
+    },
+    {
+        id: 10, src: `${process.env.PUBLIC_URL}/images/commonPhoto/cottage.jpg`, alt: 'Будиночек'
+    },
+    {
+        id: 11, src: `${process.env.PUBLIC_URL}/images/commonPhoto/cottage2_1.jpg`, alt: 'Будиночек'
+    },
+    {
+        id: 12, src: `${process.env.PUBLIC_URL}/images/commonPhoto/cottage3_1.jpg`, alt: 'Будиночек'
+    },
+    {
+        id: 13, src: `${process.env.PUBLIC_URL}/images/commonPhoto/cottage4_1.jpg`, alt: 'Будиночек'
+    },
+    {
+        id: 14, src: `${process.env.PUBLIC_URL}/images/commonPhoto/fishing.jpg`, alt: 'Риболовля'
+    },
+    {
+        id: 15, src: `${process.env.PUBLIC_URL}/images/commonPhoto/fishing_gear.jpg`, alt: 'Риболовля, снасті'
+    },
+    {
+        id: 16, src: `${process.env.PUBLIC_URL}/images/commonPhoto/grill.jpg`, alt: 'Мангал'
+    },
+    {
+        id: 17, src: `${process.env.PUBLIC_URL}/images/commonPhoto/kitchen.jpg`, alt: 'Кухня'
+    },
+    {
+        id: 18, src: `${process.env.PUBLIC_URL}/images/commonPhoto/pier.jpg`, alt: 'Пірс'
+    },
+    {
+        id: 19, src: `${process.env.PUBLIC_URL}/images/commonPhoto/pier_sunset.jpg`, alt: 'Пірс, схід сонця'
+    },
+    {
+        id: 20, src: `${process.env.PUBLIC_URL}/images/commonPhoto/room.jpg`, alt: 'Кімната'
+    },
+    {
+        id: 21, src: `${process.env.PUBLIC_URL}/images/commonPhoto/view.jpg`, alt: 'Вид з будиночку'
+    },
+]
+
+export default photoDataCommon;
